Data_Analysis_AI_Agent/requirements.txt
phidata==2.7.3
streamlit==1.41.1
openai==1.58.1
duckdb==1.1.3
pandas
numpy==1.26.4